IPLS4OC6 Fix C and D
C was not quite efficient enough. I was computing the same gcd many times. So
now I compute the gcd once, and use it until I can't anymore. Then I compute
the new gcd. It's ten times as fast now on Codeforces test case #13, which
previously received TLE.

D was wrong because I misread the bounds; n can be as large as 2*10^5.