Fixed Regionals C
I previously had one vertex in the network for each cell in the grid.
Each edge had capacity equal to the cost of building a barricade at its
destination. But this doesn't work, because this means that the police
have to barricade each "side" of each cell independently - they pay
twice if they need to barricade two sides of a cell.

Now, I have two vertices in the network for each cell in the grid: one
"in" vertex and one "out" vertex. From each "in" vertex is a single edge
to its corresponding "out" vertex with capacity equal to the cost at
that cell, or infinity if the cell is '.' or 'B'. Then other edges go
from a cell's "out" vertex to its neighbor's "in" vertex, with infinite
capacity.